Upgrade dependency, update create employee validation, change id type to uint64

See merge request hcm/backend/svc-insani-go!78

+++ modules/v1/personal/model/personal.go
@@ -2,7 +2,7 @@ package model
 
 type PersonalDataPribadi struct {
 	Id            uint64 `json:"-" gorm:"primaryKey"`
-	IdKafka       int    `json:"-"`
+	IdKafka       uint64 `json:"-"`
 	NamaLengkap   string `json:"nama_lengkap"`
 	GelarDepan    string `json:"gelar_depan"`
 	GelarBelakang string `json:"gelar_belakang"`
@@ -16,18 +16,18 @@ type PersonalDataPribadiResponse struct {
 }
 
 type PersonalDataPribadiId struct {
-	Id                 int    `json:"id" gorm:"primaryKey"`
+	Id                 uint64 `json:"id" gorm:"primaryKey"`
 	NamaLengkap        string `json:"nama_lengkap"`
 	NikKtp             string `json:"nik_ktp"`
 	NikPegawai         string `json:"nik_pegawai"`
 	TglLahir           string `json:"tgl_lahir"`
 	TempatLahir        string `json:"tempat_lahir"`
-	IdAgama            int    `json:"id_agama"`
+	IdAgama            uint64 `json:"id_agama"`
 	KdAgama            string `json:"kd_agama"`
 	JenisKelamin       string `json:"jenis_kelamin"`
-	IdGolonganDarah    int    `json:"id_golongan_darah"`
+	IdGolonganDarah    uint64 `json:"id_golongan_darah"`
 	KdGolonganDarah    string `json:"kd_golongan_darah"`
-	IdStatusPernikahan int    `json:"id_status_pernikahan"`
+	IdStatusPernikahan uint64 `json:"id_status_pernikahan"`
 }
 
 func (*PersonalDataPribadiId) TableName() string {
